chore: relocated InfoBox to the shared widgets folder
